make the path::get function into method Entry::path_get
this is just code motion, but it puts like with like.
