+++ crates/wasi-common/src/lib.rs
@@ -27,7 +27,6 @@ mod fdpool;
 pub mod fs;
 mod handle;
 pub mod old;
-mod path;
 mod sandboxed_tty_writer;
 pub(crate) mod sched;
 pub mod snapshots;
